Promotes the use of OnlinePlayground identities in the Ditto docs.
With the upcoming release of https://github.com/getditto/ditto/issues/4976 the docs have been updated to promote the use of OnlinePlayground identities, instead of the prior default OfflinePlayground identities, that were requiring a call to  as part of most configurations steps.

Using the online playground mode in the code examples has cause the  call to be removed from all of them.

This should only be released to the documentation website once the new SDK version with the  API is release.